Fix too narrow 'include within' rule in aop.xml
This is the main fix for the problem stated in
https://stackoverflow.com/q/70388073/1082681.

Background: Aspect weaving is falsely limited to only target classes,
the aspect class itself was excluded. The effect was that the weaver
never finished the aspect itself - remember, it was not compiled by AJC
before. This caused "NoSuchMethodError: ...MyAspect.aspectOf()".